+++ frontend/src/components/common/SubInfo.js
@@ -15,6 +15,7 @@ const SubInfoBlock = styled.div`
     color: ${palette.gray[4]};
     padding-left: 0.25rem;
     padding-right: 0.25rem;
+    content:'\\B7';
   }
 `;
