+++ frontend/src/pages/PostListPage.js
@@ -1,69 +1,13 @@
 import React from 'react';
 import Header from '../components/common/Header';
-import styled from 'styled-components';
-import Responsive from '../components/common/Responsive';
-import Button from '../components/common/Button';
-import palette from '../lib/styles/palette';
-import SubInfo from '../components/common/SubInfo';
-const PostListBlock = styled(Responsive)`
-  margin-top: 3rem;
-`;
+import PostList from '../components/posts/PostList';
 
-const WritePostButtonWrapper = styled.div`
-  display: flex;
-  justify-content: flex-end;
-  margin-bottom: 3rem;
-`;
-
-const PostItemBlock = styled.div`
-  padding-top: 3rem;
-  padding-bottom: 3rem;
-  /* 맨 위 포스트는 padding-top 없음 */
-  &:first-child {
-    padding-top: 0;
-  }
-  & + & {
-    border-top: 1px solid ${palette.gray[2]};
-  }
-  h2 {
-    font-size: 2rem;
-    margin-bottom: 0;
-    margin-top: 0;
-    &:hover {
-      color: ${palette.gray[6]};
-    }
-  }
-  p {
-    margin-top: 2rem;
-  }
-`;
-
-const PostItem = () => {
-  return (
-    <PostItemBlock>
-      <h2>제목</h2>
-      <SubInfo username="username" publishedDate={new Date()} />
-      <p>포스트 내용의 일부분..</p>
-    </PostItemBlock>
-  );
-};
 
 const PostListPage = () => {
   return (
     <div>
       <Header />
-      <PostListBlock>
-        <WritePostButtonWrapper>
-          <Button cyan to="/write">
-            새 글 작성하기
-          </Button>
-        </WritePostButtonWrapper>
-        <div>
-          <PostItem />
-          <PostItem />
-          <PostItem />
-        </div>
-      </PostListBlock>
+      <PostList />  
     </div>
   );
 };
